Refactor gutter icon colors in own class Example to show how it works
