+++ pitmutationmate/src/main/kotlin/com/amos/pitmutationmate/pitmutationmate/visualization/HighlightGutterRenderer.kt
@@ -1,17 +1,18 @@
-import com.intellij.codeInsight.daemon.LineMarkerInfo
-import com.intellij.codeInsight.daemon.LineMarkerProvider
-import com.intellij.icons.AllIcons
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: 2023 Tim Herzig <tim.herzig@hotmail.com>
+
 import com.intellij.openapi.editor.markup.GutterIconRenderer
 import com.intellij.psi.PsiElement
+import java.awt.Color
+import java.awt.Component
+import java.awt.Graphics
 import javax.swing.Icon
 
 
-class HighlightGutterRenderer: GutterIconRenderer() {
-    val redIcon: Icon = AllIcons.Actions.Suspend
-    val greenIcon: Icon = AllIcons.Actions.StartDebugger
-    val yellowIcon: Icon = AllIcons.Actions.IntentionBulb
+class HighlightGutterRenderer(color: String): GutterIconRenderer() {
     private val toolTip = "PITest run"
     val toolTipProvider: (PsiElement) -> String = { _ -> toolTip }
+    val color: String = color
     override fun equals(other: Any?): Boolean {
         TODO("Not yet implemented")
     }
@@ -21,8 +22,62 @@ class HighlightGutterRenderer: GutterIconRenderer() {
     }
 
     override fun getIcon(): Icon {
-        return this.redIcon
+        return if (this.color == "red") {
+            RedBarIcon()
+        }else if (this.color == "green") {
+            GreenBarIcon()
+        }else {
+            YellowBarIcon()
+        }
+    }
+
+    override fun getAlignment(): Alignment {
+        return Alignment.LEFT
     }
 
 
+    private class RedBarIcon : Icon {
+        override fun paintIcon(c: Component?, g: Graphics, x: Int, y: Int) {
+            g.setColor(Color.RED)
+            g.fillRect(x, y, iconWidth, iconHeight)
+        }
+
+        override fun getIconWidth(): Int {
+            return 10
+        }
+
+        override fun getIconHeight(): Int {
+            return 16
+        }
+    }
+
+    private class GreenBarIcon : Icon {
+        override fun paintIcon(c: Component?, g: Graphics, x: Int, y: Int) {
+            g.setColor(Color.GREEN)
+            g.fillRect(x, y, iconWidth, iconHeight)
+        }
+
+        override fun getIconWidth(): Int {
+            return 10
+        }
+
+        override fun getIconHeight(): Int {
+            return 16
+        }
+    }
+
+    private class YellowBarIcon : Icon {
+        override fun paintIcon(c: Component?, g: Graphics, x: Int, y: Int) {
+            g.setColor(Color.YELLOW)
+            g.fillRect(x, y, iconWidth, iconHeight)
+        }
+
+        override fun getIconWidth(): Int {
+            return 10
+        }
+
+        override fun getIconHeight(): Int {
+            return 16
+        }
+    }
 }
